Make shifter more prominent in search
This is a really small PR that makes Shifter appear first when you start
searching for it.  According to Brandon the # titles on pages push up their
priority in the search bar.  Otherwise the first 2 things that come up have to
do with something else.  We should look at how we title pages.