kernel-helper.sh: improve environment example
The kernel-helper.sh script for jupyter notebook kernels tries to demonstrate
usage with a user-controlled conda environment. In order to make this work
properly, the environment must also be activated. Otherwise, it is just using
the selected python to start up but not other installs in that environment.

First seen by trying to use `ffmpeg` installed as binary in such an environment.